requirements.txt
asgiref==3.5.2
astroid==2.15.5
async-timeout==4.0.2
backports.zoneinfo==0.2.1
bcrypt==3.2.0
blinker==1.4
certifi==2020.6.20
cffi==1.15.1
cfgv==3.3.1
chardet==4.0.0
charset-normalizer==2.1.1
click==8.0.3
colorama==0.4.4
cryptography==3.4.8
dill==0.3.6
distlib==0.3.5
distro==1.7.0
Django==4.1.2
django-cors-headers==3.13.0
django-extensions==3.2.1
django-jet-reboot==1.3.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
fasteners==0.14.1
filelock==3.7.1
flake8==6.0.0
future==0.18.2
google-api-python-client==1.4.1
gunicorn==20.1.0
httplib2==0.20.2
identify==2.5.24
idna==2.10
importlib-metadata==4.6.4
isort==5.12.0
jeepney==0.7.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy-object-proxy==1.9.0
lockfile==0.12.2
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
mccabe==0.7.0
monotonic==1.6
more-itertools==8.10.0
nodeenv==1.8.0
oauth2client==4.1.3
oauthlib==3.2.0
olefile==0.46
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
pipdeptree==2.3.3
platformdirs==2.5.2
pre-commit==3.3.3
protobuf==3.12.4
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pycparser==2.21
pydot==1.4.2
pyflakes==3.0.1
Pygments==2.11.2
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-dateutil==2.8.1
python-decouple==3.6
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
redis==4.5.4
reportlab==3.6.8
requests==2.25.1
rsa==4.9
SecretStorage==3.3.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tomlkit==0.11.8
typing-extensions==4.6.3
uritemplate==4.1.1
urllib3==1.26.5
virtualenv==20.16.2
wadllib==1.3.6
whitenoise==6.3.0
wrapt==1.15.0
xdg==5.0.0
zipp==1.0.0
